requirements.txt
openai==0.28
flask
pyttsx3
pydub
SpeechRecognition
requests
python-dotenv
PyYAML
